refactor: extract RecordingSessionStartOutcome.swift from RecordingSessionOutcomes.swift
Splits the `RecordingSessionStartOutcome` enum (the biggest of the
three outcome enums, with success/failure/precondition cases) out of
RecordingSessionOutcomes.swift.

Byte-identical move of lines 3-10. Both files import Foundation.

RecordingSessionOutcomes.swift: 22 → 13 lines. Tests: 667.
