+++ Sources/BugNarrator/Services/RecordingSessionOutcomes.swift
@@ -1,14 +1,5 @@
 import Foundation
 
-enum RecordingSessionStartOutcome {
-    case started(RecordingSessionDraft)
-    case restored(RecordingSessionDraft)
-    case transitionInProgress
-    case busy
-    case preflightFailure(AppError)
-    case failure(Error)
-}
-
 enum RecordingSessionStopReadiness {
     case ready(RecordingSessionDraft)
     case transitionInProgress
