+++ Sources/BugNarrator/Services/RecordingSessionStartOutcome.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum RecordingSessionStartOutcome {
+    case started(RecordingSessionDraft)
+    case restored(RecordingSessionDraft)
+    case transitionInProgress
+    case busy
+    case preflightFailure(AppError)
+    case failure(Error)
+}
